Vendor deps
This is done by turning the ipfs-10-to-11 migration into its own module and vendoring its dependencies using go mod vendor.  Since this makes ipfs-10-to-11 an independent module, and no longer a package within the fs-repo-migrations module, it changes how to build it.

The top level fs-repo-migrations module must now require the ipfs-10-to-11 module in its go.mod, and then use a replace direct so that it does not actually need to download that module from github (or find it in the vendor directory).

The sharness tests must also build it as a separate module and not as a package within the fs-repo-migrations module.