11-to-12: Re-work logic to use the backup file for both migrate and revert
Before: we Query all keys and write a backup file. Then Query all keys and
perform the swaps. This may cause that our query keeps getting results for all
the keys that we are writing while we run it.

After: we perform a Prepare() step where we query all the keys and write the
backup file. Then we read the backup file to perform the swaps without making
another query.

This also means that the migration (swapper.Run()) becomes very similar to the
revert (swapper.Revert()), since both read from the backup file. We are thus
able to refactor code and re-use logic better.

We have also tuned down the noise when encountering multihashes that are not
CIDv0s.